Move utilities to their owning tool packages
Relocate battle and overworld tool helpers under their tools directories and update consumers. Refocus the import-discipline ticket on concrete ownership and dependency direction.
