+++ agent/battle/prompts.py
@@ -2,8 +2,6 @@
 
 from typing import TYPE_CHECKING
 
-from common.enums import BattleType, PokeballItem
-
 if TYPE_CHECKING:
     from agent.context import AgentContext
     from emulator.game_state import GameState
@@ -43,32 +41,3 @@ def build_battle_decision_prompt(
         state=state,
         text=initial_game_state.screen.text,
     )
-
-
-def build_battle_tool_result(
-    game_state: GameState,
-    *,
-    action_result: str,
-    dialog: str = "",
-) -> str:
-    """Build the fresh context returned by a battle tool."""
-    sections = [action_result]
-    if dialog:
-        sections.append(f'Battle dialog: "{dialog}"')
-    available_balls: list[str] = []
-    if game_state.battle.battle_type == BattleType.WILD:
-        pokeball_names = {ball.value for ball in PokeballItem}
-        available_balls = [
-            f"- {item.name} (x{item.quantity})"
-            for item in game_state.inventory.items
-            if item.name in pokeball_names
-        ]
-    sections.extend(
-        (
-            game_state.party_info,
-            "Available Poke Balls:\n" + "\n".join(available_balls) if available_balls else "",
-            game_state.battle_info,
-            "Current onscreen text:\n" + game_state.screen.text,
-        ),
-    )
-    return "\n\n".join(section for section in sections if section)
