+++ agent/battle/tools/fight/interface.py
@@ -7,7 +7,7 @@ from pydantic_ai import Tool
 
 from agent.battle.tools.errors import BattleActionUnavailableError
 from agent.battle.tools.fight.service import fight as fight_service
-from agent.battle.utils import (
+from agent.battle.tools.utils import (
     BattleToolResult,
     complete_battle_action,
     refresh_battle_observation,
