+++ agent/battle/tools/run/interface.py
@@ -6,7 +6,7 @@ from pydantic_ai import Tool
 
 from agent.battle.tools.errors import BattleActionUnavailableError
 from agent.battle.tools.run.service import run as run_service
-from agent.battle.utils import (
+from agent.battle.tools.utils import (
     BattleToolResult,
     complete_battle_action,
     refresh_battle_observation,
