+++ agent/overworld/tools/press_buttons/interface.py
@@ -8,7 +8,7 @@ from pydantic_ai import Tool
 from agent.overworld.tools.press_buttons.service import (
     press_buttons as press_buttons_service,
 )
-from agent.overworld.utils import (
+from agent.overworld.tools.utils import (
     OverworldToolResult,
     complete_overworld_action,
 )
